+++ app/src/main/java/com/codesoom/assignment/services/TaskService.java
@@ -1,47 +1,18 @@
 package com.codesoom.assignment.services;
 
 import com.codesoom.assignment.models.Task;
-import com.codesoom.assignment.repositories.TaskRepository;
-import com.codesoom.assignment.utils.IdGenerator;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
-@Service
-public class TaskService {
+public interface TaskService {
 
-    private final TaskRepository taskRepository;
-    private final IdGenerator idGenerator;
+    List<Task> getTasks();
 
-    public TaskService(TaskRepository taskRepository, IdGenerator idGenerator) {
-        this.taskRepository = taskRepository;
-        this.idGenerator = idGenerator;
-    }
+    Task addTask(Task task);
 
-    public List<Task> getTasks() {
-        return taskRepository.findAll();
-    }
+    Task getTask(Long id);
 
-    public Task addTask(Task task) {
-        Long newId = idGenerator.generateId();
-        task.setId(newId);
-        return taskRepository.save(task);
-    }
+    Task updateTask(Long id, Task newTask);
 
-    public Task getTask(Long id) {
-        Optional<Task> task = taskRepository.findOne(id);
-        return task.orElseThrow();
-    }
-
-    public Task updateTask(Long id, Task newTask) {
-        Task task = taskRepository.findOne(id).orElseThrow();
-        task.update(newTask);
-        return taskRepository.save(task);
-    }
-
-    public void deleteTask(Long id) {
-        Task task = taskRepository.findOne(id).orElseThrow();
-        taskRepository.delete(id);
-    }
+    void deleteTask(Long id);
 }
